refactor: absorb job_type into benchmark init.
