--- metriq_gym/job_type.py
@@ -1,8 +1,0 @@
-from enum import StrEnum
-
-
-class JobType(StrEnum):
-    BSEQ = "BSEQ"
-    CLOPS = "CLOPS"
-    QML_KERNEL = "QML Kernel"
-    QUANTUM_VOLUME = "Quantum Volume"
