+++ tests/test_cli.py
@@ -3,8 +3,8 @@ from tabulate import tabulate
 import pytest
 from unittest.mock import MagicMock
 from metriq_gym.cli import LIST_JOBS_HEADERS, list_jobs
+from metriq_gym.benchmarks import JobType
 from metriq_gym.job_manager import JobManager, MetriqGymJob
-from metriq_gym.job_type import JobType
 
 
 @pytest.fixture
